Check existence of directory before deleting it

+++ applications/ShapeOptimizationApplication/python_scripts/optimization_data_logger_factory.py
@@ -72,7 +72,8 @@ class optimizationDataLogger():
     # --------------------------------------------------------------------------
     def __createFolderToStoreOptimizationResults ( self ):
         resultsDirectory = self.optimizationSettings["output"]["output_directory"].GetString()
-        shutil.rmtree(resultsDirectory)
+        if os.path.exists(resultsDirectory):
+            shutil.rmtree(resultsDirectory)
         os.makedirs(resultsDirectory)
 
     # --------------------------------------------------------------------------
